fix all compiler warnings (-Wall) - a few of which were actual errors

+++ src/RooFitUnfold.cxx
@@ -159,7 +159,7 @@ Double_t RooUnfoldFunc::getValV(const RooArgSet* set) const
 RooArgList* RooUnfoldFunc::makeParameterList() const {
   //! return a list of all parameters in this function
   RooArgSet obs;
-  for(int d=0; d<this->unfolding()->response()->Hresponse()->dim(); ++d){
+  for(size_t d=0; d<this->unfolding()->response()->Hresponse()->dim(); ++d){
     obs.add(*(this->unfolding()->response()->Hresponse()->obs(d)));
   }
   RooArgSet* pset = this->getParameters(&obs);
